requirements.txt
aiogram>=3.4.0
mistralai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0